refactor(editor): extract canvas UI helpers to separate module (Refs #1505)
